add new website
old website moved to folder WEBSITE.old and the contents of the new
website is as shown in folder.
Made changes to theme, links, and more. The website hosting from Weebly
is free and included so many features such as online forms and I'm not
too sure if all the features will work once imported into GitHub.